% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R
\name{build_grid}
\alias{build_grid}
\title{Build Grid}
\usage{
build_grid(paramsCfg, thisGridLines, neqs, psUnderNull)
}
\arguments{
\item{paramsCfg, }{tibble with information about each parameter space.}

\item{thisGridLines, }{a list of the values for each parameter that will be
explored to compute test statistic in 'Stage2_rover_Perseverance()'}

\item{neqs, }{number of equations in the model (equal to the global constant
NEQS)}

\item{psUnderNull, }{number of parameters estimated under the null that want to
be stored in the Grid.}
}
\value{
a numeric grid
}
\description{
\description{This function builds a grid with columns that contain:
\itemize{
\item the values of each parameter that are desired to be explored
\item An empty value for each of the test statistics that this function looks at
S and QLLstab }
}
}
\details{
\usage{build_grid(paramsCfg, thisGridLines, neqs, psUnderNull) }
}
\examples{
#Example 1
PARAMS_CFG <- tribble(
~name,    ~iniVal, ~stepSize, ~minVal, ~maxVal, ~decimalPlaces,
'alpha',   0.50,       0.10,       0.15,    0.98,      2)  \%>\%
 dplyr::mutate('N_TICKS' = ((maxVal-minVal)/stepSize)+1)

  originsColName='iniVal'

gridLines <- build_lines(PARAMS_CFG, 'iniVal')

Grid <- build_grid(PARAMS_CFG, gridLines)

view(Grid)
}
