% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lines.R
\name{build_lines}
\alias{build_lines}
\title{Build Lines}
\usage{
build_lines(PARAMS_CFG, originsColName)
}
\arguments{
\item{PARAMS_CFG, }{tibble with information about each parameter space}

\item{originsColName, }{The column name in 'PARAMS_CFG' that contains the parameter initial values}
}
\value{
the parameter value for each line in the grid
}
\description{
\description{
This function lines of the 'build_grid()' function.
\itemize{
\item For each parameter 'build_lines()' generates a range of values from the minimum value
to the maximum value by the step size specified in PARAMS_CFG}
}
}
\details{
\usage{build_lines(PARAMS_CFG, originsColName='iniVal')}
}
\examples{
#Example 1
PARAMS_CFG <- tribble(~name, ~iniVal, ~stepSize, ~minVal, ~maxVal, ~decimalPlaces,
'beta',  0.6, 0.025, 0, 1, 1,
'gamma', 0.6, 0.025, 0, 1, 1)  \%>\%
 dplyr::mutate('N_TICKS' = ((maxVal-minVal)/stepSize)+1)
originsColName='iniVal'
thisGridLines <- build_lines(PARAMS_CFG, originsColName='iniVal')
view(thisGridLines)
}
