% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_tests.R
\name{comp_tests}
\alias{comp_tests}
\title{Comp Tests}
\usage{
comp_tests(theta0, VffLags = NULL)
}
\arguments{
\item{theta0, }{the parameters of interest}

\item{VffLags, }{amount of lags}
}
\value{
the parameter test statistics
}
\description{
\description{This function computes the test statistics}
}
\details{
\usage{function(theta0, VffLags=NULL) is used to get the test statistics}
}
