% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_global.R
\name{generate_global}
\alias{generate_global}
\title{Generation of Global Objects}
\usage{
generate_global(PARAMS_CFG, SPVALUE, ...)
}
\arguments{
\item{PARAMS_CFG, }{a tibble with information about each parameter space}

\item{SPVALUE, }{the critical value of the test statistics}

\item{Z, }{the set of instrument variables (optional, only required for the IV estimator)}

\item{X, }{the set of exogenous variables (optional)}
}
\value{
global objects
}
\description{
\description{
This function builds the global objects required to run 'Stage2_rover_Perserverence()'}
}
