% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_IV.R
\name{mom_IV}
\alias{mom_IV}
\title{Comp Tests}
\usage{
mom_IV(theta0, z)
}
\arguments{
\item{theta0, }{the parameters of interest}

\item{VffLags, }{amount of lags}
}
\value{
the parameter test statistics
}
\description{
\description{This function computes the test statistics}
}
\details{
\usage{function(theta0, VffLags=NULL) is used to get the test statistics}
}
