% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NKPC_sams300.R
\docType{data}
\name{sim_NKPC_sams300}
\alias{sim_NKPC_sams300}
\title{Inference on hybrid NKPC curve as in Mavroeidis (2005)}
\format{
An object of class \code{list} of length 2000.
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
sim_NKPC_sams300 <- readRDS("C:/Users/User/Box/iStabi/Examples/example_NKPC/sim_NKPC_sams300.rds")
usethis::use_data(sim_NKPC_sams300)
}
\usage{
sim_NKPC_sams300
}
\description{
There are 2,000 simulated models for each case each
@format ## \code{sim_NKPC_sams300}
a data frame with 4 variables and 300 observations. :
\describe{
\item{s}{ }
\item{pi}{inflation}
\item{v}{}
\item{eps}{}
...
}
}
\keyword{datasets}
