% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2_rover_Perseverance.R
\name{Stage2_rover_Perseverance}
\alias{Stage2_rover_Perseverance}
\title{Stage2_rover_Perseverance}
\usage{
Stage2_rover_Perseverance(iniVal, WTLags, localSearch = FALSE)
}
\arguments{
\item{iniVal, }{an initial value to start grid search over}

\item{WTLags, }{the amount of lags}

\item{localSearch, }{TRUE or FALSE depending if you want a global or local search respectively}
}
\value{
the parameter value not rejected by null over the created grid
}
\description{
\description{This function fills in the test statistics in the grid produced by 'build_grid().
\itemize{
\item All these variables should be defined beforehand
\item Y, Z, Ztrans: matrices with data
\item T: integer, the number of observations
\item get_b: a function which accepts the vector theta0 and outputs the value of the vector b(theta0) such that Y\%\*\%b is the moment equation.
\item Requires access to global objects: STCPARAMNAMES, ESTVALUESNAMES, SCRITICAL, NPARAMS, GET_B, nextThetas}
}
}
\details{
\usage{Stage2_rover_Perseverance(iniVal, WTLags, localSearch = FALSE  )}
}
