% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_global.R
\name{verify_global}
\alias{verify_global}
\title{Verify Global Objects
\description{
This function verifies the global objects produced by 'generate_global()'}}
\usage{
verify_global(PARAMS_CFG, Y, ...)
}
\arguments{
\item{PARAMS_CFG, }{parameters}

\item{Y, }{the set of outcomes variables}

\item{Z, }{the set of instruments}
}
\value{
a verification all objects are valid
}
\description{
Verify Global Objects
\description{
This function verifies the global objects produced by 'generate_global()'}
}
